Build out holoviews gui.

+++ spectraclass/data/base.py
@@ -175,7 +175,7 @@ class DataManager(SCSingletonConfigurable):
         return os.path.join( output_dir, f"{self.dsid()}-{self.labels_dset}.nc" )
 
     @exception_handled
-    def save_config( self ):
+    def save_config( self, overwrite = False ):
         from spectraclass.gui.spatial.map import MapManager, mm
         from spectraclass.reduction.embedding import ReductionManager, rm
         from spectraclass.features.texture.manager import TextureManager, texm
@@ -186,21 +186,22 @@ class DataManager(SCSingletonConfigurable):
         conf_dict = self.generate_config_file()
         for scope, trait_classes in conf_dict.items():
             cfg_file = os.path.realpath( self.config_file( scope, self.mode ) )
-            os.makedirs(os.path.dirname(cfg_file), 0o777, exist_ok=True)
-            lines = []
+            if overwrite or not os.path.exists(cfg_file):
+                os.makedirs(os.path.dirname(cfg_file), 0o777, exist_ok=True)
+                lines = []
 
-            for class_name, trait_map in trait_classes.items():
-                for trait_name, trait_value in trait_map.items():
-                    tval_str = f'"{trait_value}"' if isinstance(trait_value, str) else f"{trait_value}"
-                    cfg_str = f"c.{class_name}.{trait_name} = {tval_str}\n"
-                    lines.append( cfg_str )
+                for class_name, trait_map in trait_classes.items():
+                    for trait_name, trait_value in trait_map.items():
+                        tval_str = f'"{trait_value}"' if isinstance(trait_value, str) else f"{trait_value}"
+                        cfg_str = f"c.{class_name}.{trait_name} = {tval_str}\n"
+                        lines.append( cfg_str )
 
-            lgm().log(f"Writing config file: {cfg_file}")
-            with self._lock:
-                cfile_handle = open(cfg_file, "w")
-                cfile_handle.writelines(lines)
-                cfile_handle.close()
-            lgm().log(f"Config file written")
+                lgm().log(f"Writing config file: {cfg_file}")
+                with self._lock:
+                    cfile_handle = open(cfg_file, "w")
+                    cfile_handle.writelines(lines)
+                    cfile_handle.close()
+                lgm().log(f"Config file written")
 
     def generate_config_file(self) -> Dict:
         #        print( f"Generate config file, _classes = {[inst.__class__ for inst in cls.config_instances]}")
